+++ src/extensions/index.ts
@@ -1,0 +1,4 @@
+import CodeEditor from "./codemirror/CodeMirror";
+
+export { CodeEditor };
+export * from "./react-flow";